+++ src/init.c
@@ -8,15 +8,12 @@ extern SEXP is_mutable_(SEXP x_);
 extern SEXP get_refcnt_(SEXP x_);
 extern SEXP duplicate_(SEXP x_);
 
-extern SEXP ins_replace_(SEXP x_, SEXP n_, SEXP value_);
 extern SEXP ins_sort_(SEXP x_, SEXP decreasing_);
 extern SEXP ins_shuffle_(SEXP x_);
 extern SEXP ins_reverse_(SEXP x_);
-extern SEXP ins_fill_(SEXP x_, SEXP value_);
 extern SEXP ins_runif_(SEXP x_, SEXP min_, SEXP max_);
 
-extern SEXP ins_copy_(SEXP x_, SEXP y_);
-extern SEXP ins_copy_from_(SEXP x_, SEXP y_, SEXP xi_, SEXP yi_, SEXP n_);
+extern SEXP ins_copy_(SEXP x_, SEXP y_, SEXP n_, SEXP xi_, SEXP yi_);
 
 extern SEXP fmadd_(SEXP a_, SEXP b_, SEXP c_);
 extern SEXP fmsub_(SEXP a_, SEXP b_, SEXP c_);
@@ -90,12 +87,9 @@ static const R_CallMethodDef CEntries[] = {
   {"ins_sort_"   , (DL_FUNC) &ins_sort_    , 2},
   {"ins_shuffle_", (DL_FUNC) &ins_shuffle_ , 1},
   {"ins_reverse_", (DL_FUNC) &ins_reverse_ , 1},
-  {"ins_fill_"   , (DL_FUNC) &ins_fill_    , 2},
   {"ins_runif_"  , (DL_FUNC) &ins_runif_   , 3},
   
-  {"ins_replace_"  , (DL_FUNC) &ins_replace_  , 3},
-  {"ins_copy_"     , (DL_FUNC) &ins_copy_     , 2},
-  {"ins_copy_from_", (DL_FUNC) &ins_copy_from_, 5},
+  {"ins_copy_"    , (DL_FUNC) &ins_copy_    , 5},
   
   {"fmadd_" , (DL_FUNC) &fmadd_ , 3},
   {"fmsub_" , (DL_FUNC) &fmsub_ , 3},